Wait for subnamespaceanchor status = ok on create
Creating an org creates a subnamespaceanchor which creates a namespace.
The anchor's status is updated once the namespace is created. We should
wait for that status to be 'ok' so that when an org is created via the
API, the namespace will be ready when the call returns.

Issue: https://github.com/cloudfoundry/cf-k8s-api/issues/172