Run tests in container for GitHub Actions
We want to minimise the time it takes to run these tests. Because of
this we decided to installed binaries conditionally (if they are not
installed). This will retain the current behaviour for local
development. For GitHub actions all the tools will already be installed
in the ci image used for running tests.

Co-authored-by: Mario Nitchev <marionitchev@gmail.com>