Use leases to prevent duplicate org names
We create a lease in the system namespace with a deterministic name
derived from the sha of the resource type, namespace and name. If the
lease already exists with than name, we fail with a duplicate name
error. If we can create the lease, we proceed with creating the
resource.

Leases are automatically garbage collected on deletion of the resource
due to setting the owner ref.

We have not shown update here. The procedure would be to attempt to
create a lease using the new name. If that succeeds, procede with the
rename. The old lease should be manually removed.

Co-authored-by: Kieron Browne <kbrowne@vmware.com>